requirements.txt
Django==1.7.3
Pillow==2.8.1
django-angular==0.7.13
django-autofixture==0.9.2
django-filter==0.9.2
djangorestframework==3.1.1
ipython==3.1.0
pep8==1.6.2
psycopg2==2.6
